Color in ray tracer is now based on float type of its components.

+++ src/Light.hs
@@ -6,7 +6,7 @@ import Color
 import Util
 
 -- TODO: use light color in calculations
-data Light = Directional Vector3D Float (Color Int)
+data Light = Directional Vector3D Float (Color Float)
 
 lightDir :: Light -> Vector3D
 lightDir (Directional dir _ _) = dir
@@ -14,5 +14,5 @@ lightDir (Directional dir _ _) = dir
 lightIntensity :: Light -> Float
 lightIntensity (Directional _ intensity _) = intensity
 
-lightColor :: Light -> Color Int
+lightColor :: Light -> Color Float
 lightColor (Directional _ _ color) = color
