+++ src/Material.hs
@@ -2,8 +2,8 @@ module Material where
 
 import Color
 
-data Material = DiffusiveMaterial (Color Int) | DiffusiveAndSpecularMaterial (Color Int) Int
+data Material = DiffusiveMaterial (Color Float) | DiffusiveAndSpecularMaterial (Color Float) Int
 
-materialColor :: Material -> Color Int
+materialColor :: Material -> Color Float
 materialColor (DiffusiveMaterial c) = c
 materialColor (DiffusiveAndSpecularMaterial c _) = c
